Add more flexible session connect/create mechanism.
This change exposes some new mechanisms as public apis that can
be used to build new types of connection pools and load balancer
strategies outside the cdrs-tokio crate:
- Exposes the generic_connection_pool::ConnectionPool trait
outside the crate.
- Adds a trait for user defined connection configuration and
a connect() function that can use it (in future, the other
convenience connect_* functions could be implemented in terms of
this and the N:M proliferation of connection functions might be
avoidable).

This allows the client to be in control of connection creation,
potentially in concert with their load balancing discovery strategy,
in case they have unusual needs where the two interact (ie.
a particular tls library, an alternative proprietary node discovery
mechanism, or something else).