Remove unused (and even incorrect) CDRSConnection trait methods.
These functions are not used anywhere in the cdrs codebase, so may be
vestigal. The implementations of try_clone in particular also do not
match the docstring in the trait, which claims that it should clone
the *same underlying stream* but instead it creates a whole new connection,
also necessitating carrying around a bunch of configuration baggage
for the life of the connection that's, in practice, never used.