Concentrate important types into associated types
This change removes large swaths of kind of pointless type
passing busywork from all over the system, concentrating the
information into a couple of key associated types that
are available from everywhere needed.

This greatly simplifies a lot of the code base, and makes
it easier to reason about the types involved. It likely
also makes error messages more readable.