Refactor ConnectionConfig -> GenericClusterConfig
- Rename to reflect that it's the configuration of the cluster
as a whole rather than an individual connection configuration.
- Add an associated type for the 'address' of nodes to allow
for including more information than just a SocketAddr where
needed (eg. hostname for TLS host validation or other per-
connection parameters).
- Return a ConnectionPool instead of a bb8::ConnectionManager
because that allows for the connect function to properly
translate the address given to it into a flat IP:port for
ConnectionPool's consumption, and also just simplifies
the type being returned.